data-tools/ingest/requirements.txt
configparser==3.5.0
netCDF4==1.3.1
numpy==1.13.3
pandas==0.21.0
psycopg2==2.7.3.2
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
scipy==1.0.0
six==1.11.0
xarray==0.10.0
